Add comments, cleanup
