+++ Cargo.toml
@@ -41,7 +41,6 @@ p3-matrix = { git = "https://github.com/Plonky3/Plonky3" }
 p3-merkle-tree = { git = "https://github.com/Plonky3/Plonky3" }
 p3-symmetric = { git = "https://github.com/Plonky3/Plonky3" }
 p3-uni-stark = { git = "https://github.com/Plonky3/Plonky3" }
-p3-util = { git = "https://github.com/Plonky3/Plonky3" }
 
 # Other common dependencies
 itertools = { version = "0.14.0", default-features = false, features = [
