+++ circuit/Cargo.toml
@@ -12,7 +12,9 @@ categories.workspace = true
 [dependencies]
 
 # Plonky3 dependencies
+p3-air.workspace = true
 p3-field.workspace = true
+p3-matrix.workspace = true
 p3-uni-stark.workspace = true
 
 # Other common dependencies
