+++ recursion/Cargo.toml
@@ -11,6 +11,7 @@ categories.workspace = true
 
 [dependencies]
 # Plonky3 dependencies
+p3-air.workspace = true
 p3-commit.workspace = true
 p3-field.workspace = true
 p3-fri.workspace = true
@@ -26,10 +27,7 @@ serde.workspace = true
 p3-circuit.workspace = true
 
 [dev-dependencies]
-p3-air.workspace = true
 p3-baby-bear.workspace = true
 p3-challenger.workspace = true
 p3-dft.workspace = true
-p3-matrix.workspace = true
-p3-util.workspace = true
 rand.workspace = true
