+++ recursion/src/lib.rs
@@ -4,3 +4,4 @@ extern crate alloc;
 
 pub mod recursive_pcs;
 pub mod recursive_traits;
+mod test_utils;
